Move FormActionTests back to test_models.py
